bug: DummyNode pubkey invalid utf8

+++ block/src/block.rs
@@ -238,13 +238,13 @@ impl Verifiable for Block {
         }
 
         if let Some((hash, pointers)) =
-            dependencies.0.get_lowest_pointer(self.header.block_nonce as u128)
+            dependencies.0.get_lowest_pointer(self.header.block_nonce as u64)
         {
             if hash == self.header.claim.hash {
                 if let Some(claim_pointer) = self
                     .header
                     .claim
-                    .get_pointer(self.header.block_nonce as u128)
+                    .get_pointer(self.header.block_nonce as u64)
                 {
                     if pointers != claim_pointer {
                         return Err(Self::Error {
